+++ src/models/StyledNativeComponent.js
@@ -7,6 +7,7 @@ import type { Theme } from './ThemeProvider'
 import isTag from '../utils/isTag'
 import isStyledComponent from '../utils/isStyledComponent'
 import getComponentName from '../utils/getComponentName'
+import determineTheme from '../utils/determineTheme'
 import type { RuleSet, Target } from '../types'
 
 import { CHANNEL, CHANNEL_NEXT, CONTEXT_CHANNEL_SHAPE } from './ThemeProvider'
@@ -32,7 +33,6 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
         this.context[CHANNEL_NEXT].unsubscribe(this.unsubscribeId)
       }
     }
-
 
     buildExecutionContext(theme: any, props: any) {
       const { attrs } = this.constructor
@@ -67,15 +67,9 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
         const { subscribe } = styledContext
         this.unsubscribeId = subscribe(nextTheme => {
           // This will be called once immediately
-
-          // Props should take precedence over ThemeProvider, which should take precedence over
-          // defaultProps, but React automatically puts defaultProps on props.
-          const { defaultProps } = this.constructor
-          /* eslint-disable react/prop-types */
-          const isDefaultTheme = defaultProps && this.props.theme === defaultProps.theme
-          const theme = this.props.theme && !isDefaultTheme ? this.props.theme : nextTheme
-          /* eslint-enable */
+          const theme = determineTheme(this.props, nextTheme, this.constructor.defaultProps)
           const generatedStyles = this.generateAndInjectStyles(theme, this.props)
+
           this.setState({ theme, generatedStyles })
         })
       } else {
@@ -91,13 +85,7 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
 
     componentWillReceiveProps(nextProps: { theme?: Theme, [key: string]: any }) {
       this.setState((oldState) => {
-        // Props should take precedence over ThemeProvider, which should take precedence over
-        // defaultProps, but React automatically puts defaultProps on props.
-        const { defaultProps } = this.constructor
-        /* eslint-disable react/prop-types */
-        const isDefaultTheme = defaultProps && nextProps.theme === defaultProps.theme
-        const theme = nextProps.theme && !isDefaultTheme ? nextProps.theme : oldState.theme
-        /* eslint-enable */
+        const theme = determineTheme(nextProps, oldState.theme, this.constructor.defaultProps)
         const generatedStyles = this.generateAndInjectStyles(theme, nextProps)
 
         return { theme, generatedStyles }