+++ CHANGELOG.md
@@ -6,7 +6,26 @@ All notable changes to this project will be documented in this file. If a contri
 
 ## Unreleased
 
+- Add basic support for style objects in the Typescript definitions (see [#1123](https://github.com/styled-components/styled-components/pull/1123))
 - Fix ref warning using withTheme HOC and stateless function components, thanks to [@MatthieuLemoine](https://github.com/MatthieuLemoine) (see [#1205](https://github.com/styled-components/styled-components/pull/1205))
+
+## [v2.2.4] - 2017-11-29
+
+- Disable static styles caching when hot module replacement is enabled.
+- Bump minimum `stylis` version to 3.4.0, adjust the semver caret target (see [#1296](https://github.com/styled-components/styled-components/pull/1296))
+
+## [v2.2.3] - 2017-10-29
+
+- Fix incorrect StyledNativeComponent#componentWillReceiveProps implementation (see [#1276](https://github.com/styled-components/styled-components/pull/1276))
+
+## [v2.2.2] - 2017-10-24
+
+- Prevent `withTheme` HOC from breaking when passing a theme from `defaultProps`, thanks to [@kutyel](https://github.com/kutyel) (see [#1130](https://github.com/styled-components/styled-components/pull/1130))
+- Refactor out theme logic in StyledComponent's componentWillMount & componentWillReceiveProps (see [#1130](https://github.com/styled-components/styled-components/issues/1130))
+- Add onReset to valid react props list (see [#1234](https://github.com/styled-components/styled-components/pull/1234))
+- Add support for ServerStyleSheet PropType in both StyleSheetManager and StyledComponent (see [#1245](https://github.com/styled-components/styled-components/pull/1245))
+- Prevent component styles from being static if `attrs` are dynmaic (see [#1219](https://github.com/styled-components/styled-components/pull/1219))
+- Changed 'too many classes' error to recommend attrs for frequently changed styles (see [#1213](https://github.com/styled-components/styled-components/pull/1213))
 
 ## [v2.2.1] - 2017-10-04
 
@@ -286,7 +305,10 @@ All notable changes to this project will be documented in this file. If a contri
 
 - Fixed compatibility with other react-broadcast-based systems (like `react-router` v4)
 
-[Unreleased]: https://github.com/styled-components/styled-components/compare/v2.2.1...master
+[Unreleased]: https://github.com/styled-components/styled-components/compare/v2.2.4...master
+[v2.2.4]: https://github.com/styled-components/styled-components/compare/v2.2.3...v2.2.4
+[v2.2.3]: https://github.com/styled-components/styled-components/compare/v2.2.2...v2.2.3
+[v2.2.2]: https://github.com/styled-components/styled-components/compare/v2.2.1...v2.2.2
 [v2.2.1]: https://github.com/styled-components/styled-components/compare/v2.2.0...v2.2.1
 [v2.2.0]: https://github.com/styled-components/styled-components/compare/v2.1.1...v2.2.0
 [v2.1.2]: https://github.com/styled-components/styled-components/compare/v2.1.1...v2.1.2